refactored version

+++ src/main/java/com/example/bilingualb8/services/questions/impl/TypeWhatYouHearQuestionServiceImpl.java
@@ -1,4 +1,5 @@
-package com.example.bilingualb8.services.impl.questionImpl;
+package com.example.bilingualb8.services.questions.impl;
+
 import com.example.bilingualb8.dto.requests.questions.type_what_you_hear.TypeWhatYouHearQuestionRequest;
 import com.example.bilingualb8.dto.requests.questions.type_what_you_hear.TypeWhatYouHearQuestionUpdateRequest;
 import com.example.bilingualb8.dto.responses.SimpleResponse;
@@ -6,7 +7,6 @@ import com.example.bilingualb8.dto.responses.questions.type_what_you_hear.TypeWh
 import com.example.bilingualb8.entity.File;
 import com.example.bilingualb8.entity.Question;
 import com.example.bilingualb8.entity.Test;
-import com.example.bilingualb8.enums.QuestionType;
 import com.example.bilingualb8.exceptions.NotFoundException;
 import com.example.bilingualb8.repositories.QuestionRepository;
 import com.example.bilingualb8.repositories.TestRepository;
@@ -14,7 +14,6 @@ import com.example.bilingualb8.repositories.custom.CustomTypeWhatYouHearQuestion
 import com.example.bilingualb8.services.questions.TypeWhatYouHearQuestionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -28,12 +27,15 @@ public class TypeWhatYouHearQuestionServiceImpl implements TypeWhatYouHearQuesti
 
     @Override
     public SimpleResponse saveTypeWhatYouHearQuestion(TypeWhatYouHearQuestionRequest request) {
-        Test test = testRepository.findById(request.getTestId()).orElseThrow(() -> new NoSuchElementException(String.format("Test with id = : %d not found")));
+        Test test = testRepository.findById(request.getTestId()).orElseThrow(() ->
+                new NoSuchElementException(String.format("Test with id  %d was not found", request.getTestId())));
         Question question = Question.builder()
                 .title(request.getTitle())
                 .duration(request.getDuration())
                 .questionType(request.getQuestionType())
                 .numberOfReplays(request.getNumberOfReplays())
+                .correctAnswer(request.getCorrectAnswer())
+                .questionOrder(request.getQuestionOrder())
                 .test(test)
                 .build();
         File file = new File(request.getFileRequest().getFileType(), request.getFileRequest().getFileUrl(), question);
@@ -49,14 +51,21 @@ public class TypeWhatYouHearQuestionServiceImpl implements TypeWhatYouHearQuesti
 
     @Override
     public TypeWhatYouHearQuestionResponse getTypeWhatYouHearQuestionById(Long id) {
-        return customTypeWhatYouHearQuestionRepository.getTypeWhatYouHearQuestionById(id).orElseThrow(()-> new
-          NoSuchElementException(String.format("Question  with id: %s not found!")));
+        return customTypeWhatYouHearQuestionRepository.getTypeWhatYouHearQuestionById(id).orElseThrow(() -> new
+                NoSuchElementException(String.format("Question  with id: %s not found!", id)));
     }
 
     @Override
     public SimpleResponse deleteTypeWhatYouHearById(Long id) {
+        if (!questionRepository.existsById(id)) {
+            return SimpleResponse.builder()
+                    .message(String.format("Question with id : %s doesn't exist !", id))
+                    .build();
+        }
         questionRepository.deleteById(id);
-        return SimpleResponse.builder().message(String.format("Question with id: %s successfully deleted!")).build();
+        return SimpleResponse.builder()
+                .message(String.format("Question with id : %s successfully deleted !", id))
+                .build();
     }
 
     @Override
@@ -68,13 +77,13 @@ public class TypeWhatYouHearQuestionServiceImpl implements TypeWhatYouHearQuesti
         question.setNumberOfReplays(updateQuestionRequest.getNumberOfReplays() != null ? updateQuestionRequest.getNumberOfReplays() : question.getNumberOfReplays());
         question.setCorrectAnswer(updateQuestionRequest.getCorrectAnswer() != null ? updateQuestionRequest.getCorrectAnswer() : question.getCorrectAnswer());
 
-        if(updateQuestionRequest.getFile() != null){
+        if (updateQuestionRequest.getFile() != null) {
             File file = question.getFiles().get(0);
             file.setFileUrl(updateQuestionRequest.getFile().getFileUrl());
             file.setFileType(updateQuestionRequest.getFile().getFileType());
             question.setFiles(new ArrayList<>(List.of(file)));
         }
         questionRepository.save(question);
-        return SimpleResponse.builder().message(String.format("question witg id: %s successfully updated!", id)).build();
+        return SimpleResponse.builder().message(String.format("question with id: %s successfully updated!", id)).build();
     }
 }
